scrapers/pyproject.toml
[project]
name = "scrapers"
version = "0.1.0"
description = "ER Data scraping module that retrieves and standardizes historical and weekly data taken from sensors deployed in the environment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.11",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
